refactor: move rotate_encryption_key into app/scripts (one-off CLI convention)
It's an operational one-off run by hand, matching the existing app/scripts/
modules (reset_wiki_to_seed, seed_onboarding). Invoke as
python -m app.scripts.rotate_encryption_key. No logic change.
