+++ .env.example
@@ -10,7 +10,7 @@ DEV_MODE=true
 # Optional dedicated key for at-rest column encryption. Unset = derive the
 # encryption key from SECRET_KEY (the default). Set it to rotate the encryption
 # key independently of the cookie-signing key; when you change it, re-encrypt
-# existing data with `OLD_ENCRYPTION_KEY_SECRET=<prev> python -m app.db.rotate_encryption_key`.
+# existing data with `OLD_ENCRYPTION_KEY_SECRET=<prev> python -m app.scripts.rotate_encryption_key`.
 # ENCRYPTION_KEY_SECRET=
 
 # Where the wiki git repo lives. Relative paths are resolved against the
